fix parsing precursor charge in MGF files
This fixes two issues:
- the charge value in the PEPMASS header can have the same suffixed
  format as the CHARGE header
- the order of the headers isn't determined, so the charge value should
  still be saved correctly if the CHARGE header is before the PEPMASS
  header

NOTE: I'm not sure if there is a clear precedence between PEPMASS or
CHARGE here, but it might be better to only do this if charge isn't set
already, to keep the behaviour consistent regardless of order. There
could be multiple PEPMASS headers, but it's not clear to me how the
CHARGE header should behave then: should there be multiple of them, or
should we actually set the charge of all precursor ions, like we do now
if CHARGE is first.