style: format, migrate to 2024 Edition, and fix remaining clippy lints (#31)
* style: run (patched) rustfmt

I was originally perplexed as to why the crate wasn't already formatted,
but quickly found out why... Though the PR seems to have languished for
some time, I've run a patched version of rustfmt that completes without
crashing: https://github.com/rust-lang/rustfmt/pull/6396

* style: fix `mismatched_lifetime_syntaxes` lint

* style: remove unnecessary macro imports

This is a strange case — I had to look this up myself! Because you have
a `#[macro_use]` above the `params` module in `lib.rs`, all of the
macros in `params.rs` are automatically made available crate-wide. This
import of the macro in `scan_properties.rs` is therefore redundant!

The same logic applies to `impl_param_described`.

https://lukaswirth.dev/tlborm/decl-macros/minutiae/scoping.html

* style: fix clippy `derivable_impls`

For `enum`s, you can add a tag and `#[derive(Default)]` instead of
manually implementing the same code:
https://doc.rust-lang.org/std/default/trait.Default.html#enums

* style: fix clippy `single_match`

* style: fix clippy `redundant_closure`

* style: clippy fix `unnecessary_unwrap`

* style: clippy fix `manual_is_multiple_of`

NOTE: This method was added in a recent version of Rust (1.87.0), so
changing to this code means that our minimum supported Rust version
would become that! If you'd rather keep compatibility with
older-than-stable versions of Rust, we should leave this unchanged!

* chore(edition): run `cargo update`

* style(edition): run `cargo --fix edition`

Performs some automatic migration to the 2024 edition of Rust. Will be
followed by a cleanup commit.

* style(edition): bump to Rust 2024

* style(edition): reformat using (patched) rustfmt

This is now a different, newer PR that supports the 2024 edition:
https://github.com/rust-lang/rustfmt/pull/6630

* style(edition): `expr_2021` -> `expr` in macros

The 2024 edition changed `expr`s to accept `const` and `_` expressions,
but because no macros in `mzdata` included any `const` or `_` keywords
in their syntax, no change needs to be made.

* style(edition): `match` back to `if let`

None of the migrated code seemed to depend on any exact `Drop` timing,
so the overly-cautious `if let` -> `match` migration was unnecessary.

* style(edition): change `gen` to `this_generation`

In the 2024 edition, `gen` became a reserved Rust keyword. Using `r#gen`
is still possible, but probably not best practice, so I've renamed the
old `gen` variables.

* style: fix clippy `let_and_return`

* style: fix clippy `collapsible_if`